--- packages/components/src/FormatFile/InternalThumbnailImage.module.css.d.ts
@@ -1,7 +1,0 @@
-declare const styles: {
-  readonly "image": string;
-  readonly "fadeIn": string;
-  readonly "hidden": string;
-};
-export = styles;
-
