+++ packages/components/src/Thumbnail/Thumbnail.module.css
@@ -1,3 +1,24 @@
+.image {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  animation: fadeIn 200ms ease;
+}
+
+.hidden {
+  display: none;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+
+  to {
+    opacity: 1;
+  }
+}
+
 .content {
   display: grid;
   grid-template-columns: auto;
