+++ packages/components/src/Thumbnail/Thumbnail.module.css.d.ts
@@ -1,4 +1,7 @@
 declare const styles: {
+  readonly "image": string;
+  readonly "fadeIn": string;
+  readonly "hidden": string;
   readonly "content": string;
   readonly "hasName": string;
   readonly "large": string;
