+++ packages/components/src/Thumbnail/types.ts
@@ -1,0 +1,7 @@
+import { FileUpload } from "../InputFile";
+
+export interface ThumbnailProps {
+  file: FileUpload;
+  size: "base" | "large";
+  readonly compact: boolean;
+}
